# docs/ARCH/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SPOrTS API",
    "version": "1.0.0",
    "description": "Contratti congelati rispetto all'implementazione attuale (core endpoints)"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "summary": "Register",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {"type": "object"} } } },
        "responses": { "201": {"description": "Created"}, "400": {"description": "Validation"}, "500": {"description": "Server"} }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {"type": "object"} } } },
        "responses": { "200": {"description": "OK"}, "401": {"description": "Invalid"}, "500": {"description": "Server"} }
      }
    },
    "/api/auth/verify": {
      "get": { "security": [{"bearerAuth": []}], "summary": "Verify token", "responses": { "200": {"description": "OK"}, "401": {"description": "Unauthorized"} } }
    },
    "/api/auth/me": {
      "get": { "security": [{"bearerAuth": []}], "summary": "Current user", "responses": { "200": {"description": "OK"}, "401": {"description": "Unauthorized"} } }
    },
...
    "/api/venues": {
      "get": { "security": [{"bearerAuth": []}], "summary": "Tenant venues", "responses": { "200": {"description": "OK"} } },
      "post": { "security": [{"bearerAuth": []}], "summary": "Create venue", "responses": { "201": {"description": "Created"}, "400": {"description": "Validation"} } }
    },
...
    "/api/bookings": {
      "post": { "summary": "Create booking", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#\/components\/schemas\/Booking"}}}}, "responses": { "201": {"description": "Created"}, "400": {"description": "Validation"} } }
    },
...
    "/api/match-announcements/match/{matchId}/venues": {
      "get": { "summary": "Venues for a match", "parameters": [{"in":"path","name":"matchId","required":true,"schema":{"type":"string"}}], "responses": { "200": {"description": "OK"} } }
    },
...
    "/api/match-announcements": { "post": { "security": [{"bearerAuth": []}], "summary": "Create announcement", "responses": { "201": {"description": "Created"}, "400": {"description": "Validation"}, "409": {"description": "Conflict"} } } },
...
    "/api/offer-templates": { "get": { "security": [{"bearerAuth": []}], "summary": "List offer templates", "responses": { "200": {"description": "OK"} } }, "post": { "security": [{"bearerAuth": []}], "summary": "Create offer template", "responses": { "201": {"description": "Created"} } } },
...
    "/api/global-matches": {
      "get": {
        "summary": "Lista partite globali (admin view)",
        "parameters": [
          {"in":"query","name":"league","schema":{"type":"string"}},
          {"in":"query","name":"fromDate","schema":{"type":"string"}},
          {"in":"query","name":"toDate","schema":{"type":"string"}},
          {"in":"query","name":"limit","schema":{"type":"integer"}},
          {"in":"query","name":"page","schema":{"type":"integer"}}
        ],
        "responses": { "200": {"description": "OK"} }
      }
    },
...
    "/api/sync-jobs": {
      "post": { "security": [{"bearerAuth": []}], "summary": "Crea job di sync fixtures", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"league": {"type": "string"}, "dateRange": {"type": "object"}, "syncInfo": {"type": "object"}}}}}}, "responses": { "200": {"description": "OK"}, "400": {"description": "Bad request"} } },
      "get": { "security": [{"bearerAuth": []}], "summary": "Lista job utente", "responses": { "200": {"description": "OK"} } }
    },
...
    "/api/tenants": { "get": { "security": [{"bearerAuth": []}], "summary": "Lista tenants (admin)", "responses": { "200": {"description": "OK"} } } }
  }
}